+++ railties/lib/rails/cli.rb
@@ -8,6 +8,12 @@ Rails::AppRailsLoader.exec_app_rails
 require 'rails/ruby_version_check'
 Signal.trap("INT") { puts; exit(1) }
 
+if ARGV.first == 'engine'
+  ARGV.shift
+  ARGV.unshift('plugin', 'new')
+  ARGV << '--mountable'
+end
+
 if ARGV.first == 'plugin'
   ARGV.shift
   require 'rails/commands/plugin_new'